test: generalized userupdateformat function to removeNull
