+++ test/unit/userControllerUtils.test.js
@@ -1,5 +1,5 @@
-import { userPayloadFormat, getToken, updateUserInputFormat } from "../../src/utils/userControllerUtils";
-import { dbPayload, invalidCreateUserInput, userPayload } from "../utils/testValues";
+import { userPayloadFormat, getToken  } from "../../src/utils/userControllerUtils";
+import { dbPayload,  userPayload } from "../utils/testValues";
 
 describe("Unit test of userPayloadFormat()", () => {
   describe("Given database object and token to format", () => {
@@ -17,43 +17,6 @@ describe("Unit test of getToken()", () => {
       const test = getToken("Bearer token");
 
       expect(test).toEqual("token");
-    });
-  });
-  describe("Unit test of updateUserInputFormat()", () => {
-    describe("Given keys not related to user properties", () => {
-      it("should return null", () => {
-        const test = updateUserInputFormat(invalidCreateUserInput);
-
-        expect(test).toEqual(null);
-      });
-    });
-    describe("Given keys related to user properties and all values are null", () => {
-      it("should return null", () => {
-        const update = {
-          email: null,
-          username: null,
-          password: null,
-          bio: null,
-          image: null
-        }
-        const test = updateUserInputFormat(update);
-
-        expect(test).toEqual(null);
-      });
-    });
-    describe("Given keys related to user properties with at least 1 key not valued null", () => {
-      it("should return null", () => {
-        const update = {
-          email: null,
-          username: null,
-          password: null,
-          bio: "to test or not to test",
-          image: null
-        }
-        const test = updateUserInputFormat(update);
-
-        expect(test).toEqual({ bio: update.bio });
-      });
     });
   });
 });
